feat(test): make Expectations first-level class in :test-solve
