+++ solve-classic/src/commonTest/kotlin/it/unibo/tuprolog/solve/SolveClassicTest.kt
@@ -1,0 +1,11 @@
+package it.unibo.tuprolog.solve
+
+import kotlin.jvm.JvmField
+
+object SolveClassicTest {
+    @JvmField
+    val expectations = Expectations(
+        prologShouldWork = true,
+        classicShouldWork = true
+    )
+}
