+++ test-solve/src/commonMain/kotlin/it/unibo/tuprolog/solve/Expectations.kt
@@ -1,0 +1,8 @@
+package it.unibo.tuprolog.solve
+
+data class Expectations(
+    val classicShouldWork: Boolean = false,
+    val streamsShouldWork: Boolean = false,
+    val prologShouldWork: Boolean = false,
+    val problogShouldWork: Boolean = false,
+)
